+++ test/move_divider_test.dart
@@ -16,6 +16,7 @@ void main() {
                 ],
                 containerSize: 110,
                 dividerThickness: 10,
+                dividerHandleBuffer: 0,
                 sizeUnderflowPolicy: SizeUnderflowPolicy.stretchLast);
 
             helper.moveAndTest(
@@ -34,6 +35,7 @@ void main() {
                 ],
                 containerSize: 110,
                 dividerThickness: 10,
+                dividerHandleBuffer: 0,
                 sizeUnderflowPolicy: SizeUnderflowPolicy.stretchLast);
 
             helper.moveAndTest(
@@ -52,6 +54,7 @@ void main() {
                 ],
                 containerSize: 110,
                 dividerThickness: 10,
+                dividerHandleBuffer: 0,
                 sizeOverflowPolicy: SizeOverflowPolicy.shrinkLast);
 
             helper.moveAndTest(
@@ -70,6 +73,7 @@ void main() {
                 ],
                 containerSize: 110,
                 dividerThickness: 10,
+                dividerHandleBuffer: 0,
                 sizeOverflowPolicy: SizeOverflowPolicy.shrinkLast);
 
             helper.moveAndTest(
@@ -86,7 +90,8 @@ void main() {
             TestHelper helper = TestHelper(
                 areas: [Area(data: 'a', size: 200), Area(data: 'b', flex: 1)],
                 containerSize: 310,
-                dividerThickness: 10);
+                dividerThickness: 10,
+                dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: -100, pushDividers: false, rest: 0);
@@ -100,7 +105,8 @@ void main() {
             TestHelper helper = TestHelper(
                 areas: [Area(data: 'a', size: 50), Area(data: 'b', flex: 1)],
                 containerSize: 105,
-                dividerThickness: 5);
+                dividerThickness: 5,
+                dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: 200, pushDividers: false, rest: 0);
@@ -116,7 +122,8 @@ void main() {
             TestHelper helper = TestHelper(
                 areas: [Area(data: 'a', flex: 1), Area(data: 'b', flex: 1)],
                 containerSize: 110,
-                dividerThickness: 10);
+                dividerThickness: 10,
+                dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: 0, pushDividers: false, rest: 0);
@@ -130,7 +137,8 @@ void main() {
             TestHelper helper = TestHelper(
                 areas: [Area(data: 'a', flex: 1), Area(data: 'b', flex: 1)],
                 containerSize: 110,
-                dividerThickness: 10);
+                dividerThickness: 10,
+                dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: -1, pushDividers: false, rest: 0);
@@ -144,7 +152,8 @@ void main() {
             TestHelper helper = TestHelper(
                 areas: [Area(data: 'a', flex: 1), Area(data: 'b', flex: 1)],
                 containerSize: 110,
-                dividerThickness: 10);
+                dividerThickness: 10,
+                dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: 1, pushDividers: false, rest: 0);
@@ -158,7 +167,8 @@ void main() {
             TestHelper helper = TestHelper(
                 areas: [Area(data: 'a', flex: 1), Area(data: 'b', flex: 1)],
                 containerSize: 110,
-                dividerThickness: 10);
+                dividerThickness: 10,
+                dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: -200, pushDividers: false, rest: -150);
@@ -172,7 +182,8 @@ void main() {
             TestHelper helper = TestHelper(
                 areas: [Area(data: 'a', flex: 1), Area(data: 'b', flex: 1)],
                 containerSize: 110,
-                dividerThickness: 10);
+                dividerThickness: 10,
+                dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: 200, pushDividers: false, rest: 150);
@@ -191,7 +202,7 @@ void main() {
               Area(data: 'a', size: 100),
               Area(data: 'b', flex: 1),
               Area(data: 'c', flex: 1)
-            ], containerSize: 110, dividerThickness: 5);
+            ], containerSize: 110, dividerThickness: 5, dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 1, pixels: 20, pushDividers: false, rest: 0);
@@ -208,7 +219,7 @@ void main() {
               Area(data: 'a', size: 100),
               Area(data: 'b', flex: 1),
               Area(data: 'c', flex: 3)
-            ], containerSize: 510, dividerThickness: 5);
+            ], containerSize: 510, dividerThickness: 5, dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: -20, pushDividers: false, rest: 0);
@@ -227,7 +238,7 @@ void main() {
               Area(data: 'a', flex: 1),
               Area(data: 'b', size: 100),
               Area(data: 'c', flex: 3)
-            ], containerSize: 510, dividerThickness: 5);
+            ], containerSize: 510, dividerThickness: 5, dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: -20, pushDividers: false, rest: 0);
@@ -246,7 +257,7 @@ void main() {
               Area(data: 'a', size: 50),
               Area(data: 'b', size: 50),
               Area(data: 'c', flex: 1)
-            ], containerSize: 110, dividerThickness: 5);
+            ], containerSize: 110, dividerThickness: 5, dividerHandleBuffer: 0);
 
             helper.moveAndTest(
                 dividerIndex: 0, pixels: -20, pushDividers: false, rest: 0);